refactor: update eslint config to new version
