--- .eslintignore
@@ -1,10 +1,0 @@
-node_modules
-coverage
-dist
-reports
-tests/test-projects/**/cypress
-website/.cache/**/*
-website/public/**/*
-packages/arch/docs/.cache
-packages/arch/docs/public
-packages/create-keystone-5-app/templates
