--- website/.eslintrc.js
@@ -1,3 +1,0 @@
-// gatsby doesn't seem to look at eslintrcs in
-// higher directories than the site's directory
-module.exports = require('../.eslintrc');
